Minor upgrade

+++ delete.php
@@ -1,13 +1,16 @@
 <?php
 
     include('dbcon.php');
-
-    if(isset($_POST["checkbox_value"])){
-        $ids = $_POST['checkbox_value'];
+    
+    if(isset($_POST["checkbox_value"]) && is_array($_POST["checkbox_value"])) {
+        $ids = $_POST['checkbox_value'];    
         $placeholders = rtrim(str_repeat('?,', count($ids)), ',');
-        $query = "DELETE FROM employees WHERE id IN ($placeholders)";
+        $query = "DELETE FROM employees WHERE id IN ($placeholders)";    
         $statement = $connect->prepare($query);
-        $statement->execute($ids);
+        foreach($ids as $key => $id) {
+            $statement->bindValue(($key+1), $id, PDO::PARAM_INT);
+        }    
+        $statement->execute();
     }
 
 ?>
